third_party/blink/web_tests/fast/forms/form-attribute-nonexistence-form-id-expected.txt

This page ensures that <input from="X"> and <label from="X"> don't associate with any form elements when there is no <form id="X">.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

label1 label2
PASS inputShouldHaveForm.form is form
PASS inputShouldNotHaveForm.form is null
PASS labelWithoutForm.form is null
PASS labelWithInvalidForm.form is null
PASS successfullyParsed is true

TEST COMPLETE
